Moved puppet config cloning in its own section

+++ setup.sh
@@ -134,15 +134,17 @@ if ! has_application puppet ; then
   elif [ "$ID" == "ubuntu" ] ; then
     $NOOP sudo apt-get install puppet-server rubygems
   fi
+fi
 
+if [[ ! -d /etc/puppet/.git ]] ; then
   echo "Cloning puppet configuration"
   $NOOP sudo rm -rf /etc/puppet /etc/hiera.yaml
   $NOOP sudo git clone http://github.com/gildas/setup-puppetserver.git /etc/puppet
   $NOOP sudo ln -s /etc/puppet/hiera.yaml /etc/hiera.yaml
-
-  $NOOP sudo mkdir -p /var/lib/puppet/ssl
-  $NOOP sudo chown -R puppet:puppet /var/lib/puppet/client* /var/lib/puppet/lib /var/lib/puppet/ssl
 fi
+
+[[ -d /var/lib/puppet/ssl ]] || $NOOP sudo mkdir -p /var/lib/puppet/ssl
+$NOOP sudo chown -R puppet:puppet /var/lib/puppet/client* /var/lib/puppet/lib /var/lib/puppet/ssl
 
 if [[ ! -z "$(gem list --local | grep librarian-puppet)" ]] ; then
   echo "Installing librarian for puppet"
